% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_and_visualize_GVI.R
\name{calculate_and_visualize_GVI}
\alias{calculate_and_visualize_GVI}
\title{Calculate and Visualize Green View Index (GVI) from an image}
\usage{
calculate_and_visualize_GVI(image_path)
}
\arguments{
\item{image_path}{The path of the image file to be processed.}
}
\value{
A list containing the Green View Index (GVI), the segmented image, and the green pixels image.
}
\description{
This function reads an image, performs superpixel segmentation (using the SuperpixelImageSegmentation library), calculates the Green View Index (GVI),
and returns a list containing the segmented image, the green pixels image, and the calculated GVI.
}
\examples{
\dontrun{
# Example usage with an image located at the specified path
result <- calculate_and_visualize_GVI("/path/to/your/image.png")
}
}
