% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{compute_gini_bootstrap}
\alias{compute_gini_bootstrap}
\title{Independent-observation bootstrap Gini inequality index}
\usage{
compute_gini_bootstrap(data_vector, R = 100)
}
\arguments{
\item{data_vector}{Numeric vector of values to calculate Gini for.}

\item{R}{Number of bootstrap replicates.}
}
\value{
A list with the Gini value and iid bootstrap interval. It ignores
spatial autocorrelation and is unsuitable for spatial inference.
}
\description{
Independent-observation bootstrap Gini inequality index
}
