% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_data.R
\name{.way_geom_to_matrix}
\alias{.way_geom_to_matrix}
\title{Convert OSM way geometry to matrix}
\usage{
.way_geom_to_matrix(geom_list)
}
\arguments{
\item{geom_list}{List of geometry points.}
}
\value{
A matrix of coordinates.
}
\description{
Convert OSM way geometry to matrix
}
