% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_space_clustering.R
\name{green_space_clustering}
\alias{green_space_clustering}
\title{Green Space Clustering with K-Means and Tile Layer Control in Leaflet}
\usage{
green_space_clustering(green_areas_data, num_clusters)
}
\arguments{
\item{green_areas_data}{List containing green areas data (obtained from get_osm_data function or similar).}

\item{num_clusters}{Integer number of clusters to divide the green spaces into.}
}
\value{
A Leaflet map object displaying clustered green spaces with layer control for basemap tiles.
}
\description{
This function performs K-means clustering on green spaces based on their area size and visualizes the results on a Leaflet map.
Users must specify the number of clusters. The function includes a layer control for switching between different basemap tiles.
}
\examples{
\dontrun{
  # Assuming you have already obtained green_areas_data using get_osm_data or similar
  map <- green_space_clustering(green_areas_data, num_clusters = 5)
  map # to display the map
}
}
