% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{plot_multilayer_leaflet}
\alias{plot_multilayer_leaflet}
\title{Interactive Multilayer Leaflet Map with layer controls, legends, and dynamic popups}
\usage{
plot_multilayer_leaflet(priority_data, canyon_data = NULL, palette = NULL)
}
\arguments{
\item{priority_data}{A priority grid dataset.}

\item{canyon_data}{Optional street canyon dataset.}

\item{palette}{Optional color palette for priority scores.}
}
\description{
Interactive Multilayer Leaflet Map with layer controls, legends, and dynamic popups
}
