% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{plot_priority_3d_explorer}
\alias{plot_priority_3d_explorer}
\title{Generate fully interactive WebGL 3D Decision Support Explorer}
\usage{
plot_priority_3d_explorer(
  priority_data,
  output_html,
  render_type = c("auto", "hexagons", "blocks")
)
}
\arguments{
\item{priority_data}{A priority grid dataset.}

\item{output_html}{Path to save the interactive HTML dashboard.}

\item{render_type}{One of "auto", "hexagons", or "blocks".}
}
\description{
Generate fully interactive WebGL 3D Decision Support Explorer
}
