% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{plot_priority_bivariate}
\alias{plot_priority_bivariate}
\title{Create static bivariate Priority maps}
\usage{
plot_priority_bivariate(
  priority_data,
  title = "Bivariate Shading Need & Physical Opportunity Map"
)
}
\arguments{
\item{priority_data}{A priority grid dataset.}

\item{title}{Plot title.}
}
\description{
Create static bivariate Priority maps
}
