% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{plot_priority_diamond_bivariate}
\alias{plot_priority_diamond_bivariate}
\title{Custom 45-degree rotated Diamond Bivariate Decision Map}
\usage{
plot_priority_diamond_bivariate(
  priority_data,
  title = "Where Heat Mitigation Need & Planting Opportunity Coincide",
  subtitle = NULL,
  caption = NULL,
  palette = NULL
)
}
\arguments{
\item{priority_data}{A priority grid dataset.}

\item{title}{Plot title.}

\item{subtitle}{Optional plot subtitle.}

\item{caption}{Optional plot caption.}

\item{palette}{Optional 9-color bivariate palette.}
}
\description{
Custom 45-degree rotated Diamond Bivariate Decision Map
}
