% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{plot_priority_interactive}
\alias{plot_priority_interactive}
\title{Interactive Leaflet decision widget with dynamic legends}
\usage{
plot_priority_interactive(priority_data)
}
\arguments{
\item{priority_data}{A priority grid dataset.}
}
\description{
Interactive Leaflet decision widget with dynamic legends
}
