% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_green_index.R
\name{rename_duplicate_columns}
\alias{rename_duplicate_columns}
\title{Helper function to rename duplicate columns}
\usage{
rename_duplicate_columns(df)
}
\arguments{
\item{df}{A data.frame.}
}
\description{
Helper function to rename duplicate columns
}
