% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny App}
\usage{
run_app()
}
\value{
No return value, called for side effects
}
\description{
This function runs the included Shiny app. The app provides an interactive
interface to use the functions in this package. You can download OSM data,
calculate green indices, plot green index, and save green index data as a
JSON file or as a Leaflet map in an HTML file.
}
\examples{
\dontrun{
  run_app()
}
}
