% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{screen_canyon_orientation}
\alias{screen_canyon_orientation}
\alias{emulate_canyon_microclimate}
\title{Heuristic street-orientation and canopy-gap screening}
\usage{
screen_canyon_orientation(canyon_data, latitude = 46.2)

emulate_canyon_microclimate(canyon_data, latitude = 46.2)
}
\arguments{
\item{canyon_data}{A street canyon dataset returned from build_street_canyon_priority.}

\item{latitude}{Study region latitude in decimal degrees (e.g. 28.6 for New Delhi, 46.2 for Geneva).
Used only in a latitude-weighted orientation proxy. This is not a solar
position, radiation, SVF, air-temperature, or microclimate simulation.}
}
\description{
Heuristic street-orientation and canopy-gap screening
}
