% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_green_spaces.R
\name{visualize_green_spaces}
\alias{visualize_green_spaces}
\title{Visualize Green Spaces on a Leaflet Map}
\usage{
visualize_green_spaces(green_areas_data)
}
\arguments{
\item{green_areas_data}{List containing green areas data (obtained from get_osm_data function).}
}
\value{
A Leaflet map displaying green spaces with labels and a legend, with a layer control for toggling the green spaces layer.
}
\description{
This function visualizes green spaces on a Leaflet map using the green_areas_data obtained from the get_osm_data function.
Green spaces are labeled based on their tags and have different colors in the legend. Users can switch the green spaces layer on and off.
}
\examples{
\dontrun{
  # Assuming you have already obtained green_areas_data using get_osm_data
  visualize_green_spaces(green_areas_data)
}
}
