% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{assess_urban_priority_equity}
\alias{assess_urban_priority_equity}
\title{Assess equity and inequality index with bootstrap uncertainty}
\usage{
assess_urban_priority_equity(priority_data, n_bootstrap = 250)
}
\arguments{
\item{priority_data}{A priority grid dataset returned from build_urban_priority_grid.}

\item{n_bootstrap}{Number of bootstrap iterations for uncertainty estimation (default: 250).}
}
\description{
Assess equity and inequality index with bootstrap uncertainty
}
