% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{build_street_canyon_priority}
\alias{build_street_canyon_priority}
\title{Morphological Street Canyon priority scoring engine (Mathematically Rigorous MCDA)}
\usage{
build_street_canyon_priority(priority_data)
}
\arguments{
\item{priority_data}{A priority grid dataset returned from build_urban_priority_grid.}
}
\description{
Morphological Street Canyon priority scoring engine (Mathematically Rigorous MCDA)
}
