% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{build_urban_block_priority}
\alias{build_urban_block_priority}
\title{Morphological Urban Block Subdivision and priority scoring engine (Superblocks Scale)}
\usage{
build_urban_block_priority(
  priority_data,
  w_heat = 0.5,
  w_pop = 0.5,
  w_ndvi = 0.5,
  w_canopy = 0.5
)
}
\arguments{
\item{priority_data}{A priority grid dataset returned from build_urban_priority_grid.}

\item{w_heat}{Weight of Land Surface Temperature in Heat Exposure index (default: 0.50).}

\item{w_pop}{Weight of Population in Heat Exposure index (default: 0.50).}

\item{w_ndvi}{Weight of NDVI deficit in Cooling Deficit index (default: 0.50).}

\item{w_canopy}{Weight of canopy deficit in Cooling Deficit index (default: 0.50).}
}
\description{
Morphological Urban Block Subdivision and priority scoring engine (Superblocks Scale)
}
