% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_green_index.R
\name{calculate_green_index}
\alias{calculate_green_index}
\title{Calculate green index for each edge}
\usage{
calculate_green_index(osm_data, crs_code, D = 100)
}
\arguments{
\item{osm_data}{The OpenStreetMap data. This should be a list with three components:
highways, green_areas, and trees. Each component should be a spatial data frame.
You can use the osmdata package to get the required data.}

\item{crs_code}{The EPSG code for the Coordinate Reference System (CRS).}

\item{D}{The decay parameter in the decay function, default is 100.}
}
\value{
A data frame with the green index for each edge.
}
\description{
This function calculates the green index for each edge in the road network.
}
\examples{
\dontrun{
# osm_data should be a list with three components: highways, green_areas, and trees
# You can use osmdata package to get this data.
calculate_green_index(osm_data, 2056, D = 100)
}
}
