% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_green_index.R
\name{check_duplicate_columns}
\alias{check_duplicate_columns}
\title{Helper function to check for duplicate columns}
\usage{
check_duplicate_columns(df)
}
\arguments{
\item{df}{A data.frame. The input data frame to check for duplicate columns.}
}
\description{
Helper function to check for duplicate columns
}
