% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_green_index.R
\name{check_duplicate_columns}
\alias{check_duplicate_columns}
\title{Function to check for duplicate columns and print them}
\usage{
check_duplicate_columns(df)
}
\arguments{
\item{df}{Data frame to check for duplicate columns}
}
\value{
Vector of duplicate column names
}
\description{
Function to check for duplicate columns and print them
}
