% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uh_decision.R
\name{compute_gini_bootstrap}
\alias{compute_gini_bootstrap}
\title{Highly Optimized Bootstrap Gini Inequality Index}
\usage{
compute_gini_bootstrap(data_vector, R = 100)
}
\arguments{
\item{data_vector}{Numeric vector of values to calculate Gini for.}

\item{R}{Number of bootstrap replicates.}
}
\description{
Highly Optimized Bootstrap Gini Inequality Index
}
