% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_data.R
\name{get_osm_data}
\alias{get_osm_data}
\title{Download OSM Data}
\usage{
get_osm_data(location)
}
\arguments{
\item{location}{A string representing the location (e.g., "Genève, Schweiz/Suisse/Svizzera/Svizra").}
}
\value{
A list containing:
\itemize{
\item highways: an sf object with the OSM data about highways in the specified location.
\item green_areas: an sf object with the OSM data about green areas, such as parks, forests, gardens,
and nature reserves, in the specified location.
\item trees: an sf object with the OSM data about trees in the specified location.
}
}
\description{
This function downloads OpenStreetMap (OSM) data for a specified location.
The OSM data includes data about highways, green areas, and trees in the specified location.
It requires an internet connection.
}
\examples{
\dontrun{
  osm_data <- get_osm_data("Genève, Schweiz/Suisse/Svizzera/Svizra")
}
}
