% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_green_index.R
\name{plot_green_index}
\alias{plot_green_index}
\title{Plot the green index}
\usage{
plot_green_index(green_index)
}
\arguments{
\item{green_index}{A data frame containing the calculated green index values for each edge.}
}
\description{
This function plots the green index for the highway network.
}
\examples{
\dontrun{
# Generate a sample green_index data frame
green_index <- data.frame(
  green_index = runif(1000),
  geometry = rep(sf::st_sfc(sf::st_point(c(0, 0))), 1000)
)
plot_green_index(green_index)
}
}
