% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_leaflet.R
\name{save_as_leaflet}
\alias{save_as_leaflet}
\title{Save the green index data as a Leaflet map in an HTML file}
\usage{
save_as_leaflet(edges, file_path)
}
\arguments{
\item{edges}{A data frame containing the calculated green index values for each edge.}

\item{file_path}{The file path where the HTML file will be saved.}
}
\value{
No return value, called for side effects
}
\description{
This function saves the green index data as a Leaflet map in an HTML file.
}
\examples{
\dontrun{
# Assuming you have already obtained green index data
save_as_leaflet(green_index, "green_index_map.html")
}
}
